Remove unneeded Generic from DbBase
Generic should be used as a base class when the class itself needs to be
generic -- e.g. a container type could be Container(Generic[T]). In this
case, however, the intention is to indicate that clone() returns an
object of the same type as that of the object it is called on.